requirements.txt
flask
uvicorn
requests
fastapi
torchaudio
SpeechRecognition
torch
pydantic
google
google.genai
python-multipart
pydub
soundfile
python-dotenv